# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用Riot官方API获取数据进行分析、统计项目",
        "title": "ChaosMetrics API接口文档",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.9"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/ARAM": {
            "get": {
                "description": "query @version,loc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Champion Rank"
                ],
                "summary": "请求一个ARAM英雄榜",
                "parameters": [
                    {
                        "type": "string",
                        "default": "na1",
                        "description": "Region",
                        "name": "loc",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "14.1.1",
                        "description": "Version",
                        "name": "version",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/anres.ChampionBrief"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/CLASSIC": {
            "get": {
                "description": "query @version,loc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Champion Rank"
                ],
                "summary": "请求一个CLASSIC英雄榜",
                "parameters": [
                    {
                        "type": "string",
                        "default": "na1",
                        "description": "Region",
                        "name": "loc",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "14.1.1",
                        "description": "Version",
                        "name": "version",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/anres.ChampionBrief"
                            }
                        }
                    }
                }
            }
        },
...
        "/comments": {
            "post": {
                "description": "post @ChampionID,Version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "发表一个评论",
                "parameters": [
                    {
                        "description": "Post a comment @ champion,version",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.postCommentParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "delete @id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "删除一个评论",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/comments/list": {
            "get": {
                "description": "query @ChampionID,Version,start,size",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "请求一个英雄的评论",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Ahri",
                        "description": "Champion name",
                        "name": "championID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "14.1.1",
                        "description": "Version",
                        "name": "version",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/comment.CommentsDTO"
                                            }
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/user/changepasswd": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "change @passwd,newpasswd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Service"
                ],
                "summary": "更改密码",
                "parameters": [
                    {
                        "description": "login",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.changePasswordParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "anres.ChampionBrief": {
            "type": "object",
            "properties": {
                "avg_damage_dealt": {
                    "description": "场均输出占比 12%",
                    "type": "number"
                },
                "avg_dead_time": {
                    "description": "场均死亡时长",
                    "type": "number"
                },
                "ban_rate": {
                    "description": "Ban率",
                    "type": "number"
                },
                "id": {
                    "description": "英雄ID:Aatrox",
                    "type": "string"
                },
                "image": {
                    "description": "basic info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Image"
                        }
                    ]
                },
                "pick_rate": {
                    "description": "登场率 15%",
                    "type": "number"
                },
                "win_rate": {
                    "description": "statistical data",
                    "type": "number"
                }
            }
        },
...
        "comment.Author": {
            "type": "object",
            "properties": {
                "nickName": {
                    "type": "string"
                }
            }
        },
        "comment.CommentsDTO": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/comment.Author"
                },
                "championID": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "comment.postCommentParam": {
            "type": "object",
            "required": [
                "championID",
                "comment",
                "version"
            ],
            "properties": {
                "championID": {
                    "description": "Champion name",
                    "type": "string",
                    "example": "Ahri"
                },
                "comment": {
                    "type": "string",
                    "example": "so cute"
                },
                "version": {
                    "description": "Version",
                    "type": "string",
                    "example": "14.1.1"
                }
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "full": {
                    "description": "大图文件名",
                    "type": "string"
                },
                "group": {
                    "description": "图像所属组",
                    "type": "string"
                },
                "h": {
                    "description": "图像高度",
                    "type": "integer"
                },
                "sprite": {
                    "description": "小图文件名",
                    "type": "string"
                },
                "w": {
                    "description": "图像宽度",
                    "type": "integer"
                },
                "x": {
                    "description": "图像 X 坐标",
                    "type": "integer"
                },
                "y": {
                    "description": "图像 Y 坐标",
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "user.changePasswordParam": {
            "type": "object",
            "required": [
                "newPassword",
                "password"
            ],
            "properties": {
                "newPassword": {
                    "description": "新密码",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}